requirements.txt
python-telegram-bot>=20.0
Flask
python-dotenv
redis
gunicorn
google-generativeai